+++ Sources/Sentry/SentryEnvelope.m
@@ -29,29 +29,29 @@ NS_ASSUME_NONNULL_BEGIN
 
 - (instancetype)initWithId:(SentryId *_Nullable)eventId andSdkInfo:(SentrySdkInfo *_Nullable)sdkInfo
 {
-    return [self initWithId:eventId sdkInfo:sdkInfo traceState:nil];
+    return [self initWithId:eventId sdkInfo:sdkInfo traceContext:nil];
 }
 
 - (instancetype)initWithId:(nullable SentryId *)eventId
-                traceState:(nullable SentryTraceState *)traceState
+              traceContext:(nullable SentryTraceContext *)traceContext
 {
     SentrySdkInfo *sdkInfo = [[SentrySdkInfo alloc] initWithName:SentryMeta.sdkName
                                                       andVersion:SentryMeta.versionString];
 
-    self = [self initWithId:eventId sdkInfo:sdkInfo traceState:traceState];
+    self = [self initWithId:eventId sdkInfo:sdkInfo traceContext:traceContext];
 
     return self;
 }
 
 - (instancetype)initWithId:(nullable SentryId *)eventId
                    sdkInfo:(nullable SentrySdkInfo *)sdkInfo
-                traceState:(nullable SentryTraceState *)traceState
+              traceContext:(nullable SentryTraceContext *)traceContext
 {
 
     if (self = [super init]) {
         _eventId = eventId;
         _sdkInfo = sdkInfo;
-        _traceState = traceState;
+        _traceContext = traceContext;
     }
 
     return self;