+++ Sources/Sentry/include/SentryDependencyContainer.h
@@ -2,8 +2,8 @@
 #import "SentryRandom.h"
 #import <Foundation/Foundation.h>
 
-@class SentryAppStateManager, SentryCrashWrapper, SentryThreadWrapper, SentryDispatchQueueWrapper,
-    SentrySwizzleWrapper, SentryDebugImageProvider;
+@class SentryAppStateManager, SentryCrashWrapper, SentryThreadWrapper, SentrySwizzleWrapper,
+    SentryDispatchQueueWrapper, SentryDebugImageProvider, SentryANRTracker;
 
 #if SENTRY_HAS_UIKIT
 @class SentryScreenshot, SentryUIApplication;
@@ -28,11 +28,14 @@ SENTRY_NO_INIT
 @property (nonatomic, strong) SentrySwizzleWrapper *swizzleWrapper;
 @property (nonatomic, strong) SentryDispatchQueueWrapper *dispatchQueueWrapper;
 @property (nonatomic, strong) SentryDebugImageProvider *debugImageProvider;
+@property (nonatomic, strong) SentryANRTracker *anrTracker;
 
 #if SENTRY_HAS_UIKIT
 @property (nonatomic, strong) SentryScreenshot *screenshot;
 @property (nonatomic, strong) SentryUIApplication *application;
 #endif
+
+- (SentryANRTracker *)getANRTracker:(NSTimeInterval)timeout;
 
 @end