+++ Sources/Sentry/include/SentryNetworkTracker.h
@@ -14,7 +14,8 @@ static NSString *const SENTRY_NETWORK_REQUEST_TRACKER_SPAN = @"SENTRY_NETWORK_RE
 - (void)urlSessionTaskResume:(NSURLSessionTask *)sessionTask;
 - (void)urlSessionTask:(NSURLSessionTask *)sessionTask setState:(NSURLSessionTaskState)newState;
 
-- (nullable NSDictionary *)addTraceHeader:(nullable NSDictionary *)headers;
+- (nullable NSDictionary<NSString *, NSString *> *)addTraceHeader:
+    (nullable NSDictionary<NSString *, NSString *> *)headers;
 
 - (void)enableNetworkTracking;