+++ Sources/Sentry/include/SentryTransportAdapter.h
@@ -4,7 +4,7 @@
 #import "SentryTransport.h"
 
 @class SentryEnvelope, SentryEnvelopeItem, SentryEvent, SentrySession, SentryUserFeedback,
-    SentryAttachment, SentryTraceState, SentryOptions;
+    SentryAttachment, SentryTraceContext, SentryOptions;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -27,19 +27,19 @@ SENTRY_NO_INIT
       attachments:(NSArray<SentryAttachment *> *)attachments;
 
 - (void)sendEvent:(SentryEvent *)event
-       traceState:(nullable SentryTraceState *)traceState
+     traceContext:(nullable SentryTraceContext *)traceContext
       attachments:(NSArray<SentryAttachment *> *)attachments
-    NS_SWIFT_NAME(send(event:traceState:attachments:));
+    NS_SWIFT_NAME(send(event:traceContext:attachments:));
 
 - (void)sendEvent:(SentryEvent *)event
-                 traceState:(nullable SentryTraceState *)traceState
+               traceContext:(nullable SentryTraceContext *)traceContext
                 attachments:(NSArray<SentryAttachment *> *)attachments
     additionalEnvelopeItems:(NSArray<SentryEnvelopeItem *> *)additionalEnvelopeItems
-    NS_SWIFT_NAME(send(event:traceState:attachments:additionalEnvelopeItems:));
+    NS_SWIFT_NAME(send(event:traceContext:attachments:additionalEnvelopeItems:));
 
 - (void)sendEvent:(SentryEvent *)event
       withSession:(SentrySession *)session
-       traceState:(nullable SentryTraceState *)traceState
+     traceContext:(nullable SentryTraceContext *)traceContext
       attachments:(NSArray<SentryAttachment *> *)attachments;
 
 - (void)sendUserFeedback:(SentryUserFeedback *)userFeedback NS_SWIFT_NAME(send(userFeedback:));