+++ Tests/SentryTests/SentryCrash/TestThreadInspector.swift
@@ -2,6 +2,8 @@ import Foundation
 
 class TestThreadInspector: SentryThreadInspector {
     
+    var allThreds: [Sentry.Thread]?
+    
     static var instance: TestThreadInspector {
         // We need something to pass to the super initializer, because the empty initializer has been marked unavailable.
         let inAppLogic = SentryInAppLogic(inAppIncludes: [], inAppExcludes: [])
@@ -11,7 +13,11 @@ class TestThreadInspector: SentryThreadInspector {
     }
     
     override func getCurrentThreads() -> [Sentry.Thread] {
-        return [TestData.thread]
+        return allThreds ?? [TestData.thread]
+    }
+
+    override func getCurrentThreads(withStackTrace getAllStacktraces: Bool) -> [Sentry.Thread] {
+        return allThreds ?? [TestData.thread]
     }
 
 }